Changed DB

+++ project/database/migrations/2021_05_18_220221_create_factor_de_riesgos_table.php
@@ -17,7 +17,7 @@ class CreateFactorDeRiesgosTable extends Migration
             $table->id();
             $table->integer('pregunta_id');
             $table->foreignId('beneficiario_id');
-            $table->integer('respuesta');
+            $table->float('respuesta');
             $table->text('enfermedad')->nullable();
             $table->timestamps();
